[refactor] 피드 퀘스트 CRUD 에러처리

+++ src/feeds/feeds.service.ts
@@ -13,17 +13,19 @@ export class FeedsService {
     private feedRepository: FeedRepository
   ) {}
 
-  // createFeed(createFeedDto: CreateFeedDto) {
-  //   return 'This action adds a new feed';
-  // }
-
-  async findAllFeeds() {
-    // return Feed.find();
-    const feeds = await Feed.find({ relations: ['place'] });
+  /* 모든 피드 가져오기 */
+  async findAllFeeds(): Promise<Feed[]> {
+    const feeds = await Feed.find({
+      where: {
+        deletedAt: null,
+      },
+      relations: ['place'],
+    });
     return feeds;
   }
 
-  async findOneFeed(feedId: number) {
+  /* 특정 피드 가저오기 */
+  async findOneFeed(feedId: number): Promise<Feed> {
     const content = await this.feedRepository.findOne({
       where: {
         id: feedId,
@@ -36,26 +38,29 @@ export class FeedsService {
     return content;
   }
 
-  async updateFeed(
-    feedId: number,
-    files: object[],
-    // updateFeedDto: UpdateFeedDto
-    feedContent: string
-  ) {
+  /* 피드 수정 */
+  async updateFeed(feedId: number, files: object[], feedContent: string) {
     const feed = await this.findOneFeed(feedId);
-    // const filePath = files['path'];
     const pathList = [];
     files.map((file) => {
       pathList.push(file['path']);
     });
     if (feed) {
-      // return this.feedRepository.updateFeed(feedId, pathList, updateFeedDto);
       return this.feedRepository.updateFeed(feedId, pathList, feedContent);
     }
     return `feed not found id ${feedId}`;
   }
 
-  async removeQuest(feedId: number) {
+  /* 피드 삭제 */
+  async removeQuest(feedId: number): Promise<void | object> {
+    const feed = await this.feedRepository.findOne(feedId);
+    if (!feed) {
+      // throw new NotFoundException(`feed not found`);
+      return {
+        ok: false,
+        message: `not found`,
+      };
+    }
     return this.feedRepository.deleteFeed(feedId);
   }
 }
